requirements.txt
# Core
flask>=3.0.0
gunicorn>=21.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Qiskit (Quantum Computing)
qiskit>=1.0.0
qiskit-aer>=0.14.0
qiskit-algorithms>=0.3.0
qiskit-finance>=0.4.0
qiskit-optimization>=0.6.0

# LLM & Data
google-generativeai>=0.8.0
finnhub-python>=2.4.0
yfinance>=0.2.0
pandas>=2.0.0
